refactor: 불필요 파일 정리 및 컨벤션에 맞게 store file 수정
