--- apps/user/src/stores/index.test.ts
@@ -1,7 +1,0 @@
-import { useSearchHistoryStore } from "./index";
-
-describe("stores index", () => {
-  it("useSearchHistoryStore를 export 한다", () => {
-    expect(useSearchHistoryStore).toBeDefined();
-  });
-});
